requirements.txt
asgiref
certifi
cffi
channels
charset-normalizer
cryptography
defusedxml
Django
django-allauth
django-crispy-forms
django-formtools
django-guardian
djangorestframework
djangorestframework-jwt
idna
mysqlclient
oauthlib
psycopg2-binary
pycparser
PyJWT
python3-openid
pytz
redis
requests
requests-oauthlib
sqlparse
tzdata
urllib3
whitenoise
typing_extensions
Pillow
python-decouple
gunicorn
dj-database-url